# wallet: cache IsMine scriptPubKeys to improve performance of descriptor wallets