@darosior 2023-02-06
Concept ACK. This was hit in practice by a user in #27002. From https://github.com/bitcoin/bitcoin/issues/27002#issuecomment-1412556699:
- v24.0.1:
    ```
    $ time ./bitcoin-cli -rpcwallet=desc listunspent|grep txid|wc -l
    224
    
    real	1m15.596s
    user	0m0.008s
    sys	0m0.000s
    ```
- This PR:
    ```
    $ time ./bitcoin-cli -rpcwallet=desc listunspent|grep txid|wc -l
    314

    real	0m0.464s
    user	0m0.009s
    sys	0m0.000s
    ```