@willcl-ark 2023-02-06
Concept ACK.

I wonder if the performance boost here might also be enough to close out https://github.com/bitcoin/bitcoin/issues/19942 even though it's not implementing the functionality OP there thought would solve their issue?