@achow101 2023-02-06
I've ended up rewriting a significant chunk of this PR. Instead of caching the scriptPubKeys of non-ranged descriptors, I've changed this to cache the scriptPubKeys of all descriptors. I think this makes it easier to understand and is also less fragile.

Originally I didn't cache the spks of ranged descriptors because I thought it would be difficult to update the cached spks every time a `TopUp` was called as that is used in several places internally. My solution for that issue is to have a callback in the SPKM that is called by `TopUp` to inform the containing `CWallet` of the newly added spks. This allows the spk cache to always match the descriptor's spk set.

One concern of this approach is that it maintains duplicate copies of the scriptPubKeys, thus significantly increasing the memory footprint. However, this may not really be a concern given that scriptPubKeys are relatively small and so it really just ends up being a couple of kB.

This also managed to improve the performance of `IsMine` even without a ton of non-ranged descriptors. Here are the benchmark results for the original, and the new:

Original:
|               ns/op |                op/s |    err% |          ins/op |         bra/op |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|---------------:|--------:|----------:|:----------
|           11,445.75 |           87,368.70 |    0.2% |      103,638.12 |      18,444.02 |    0.1% |      0.01 | `WalletIsMineDescriptors`
|            8,256.94 |          121,110.23 |    0.1% |       74,299.09 |      10,755.01 |    0.1% |      0.01 | `WalletIsMineLegacy`
|           12,289.10 |           81,372.91 |    0.3% |      114,209.13 |      20,204.02 |    0.1% |      0.01 | `WalletIsMineMigratedDescriptors`

New:
|               ns/op |                op/s |    err% |          ins/op |         bra/op |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|---------------:|--------:|----------:|:----------
|            1,734.82 |          576,429.48 |    0.2% |       15,191.02 |       2,579.00 |    0.0% |      0.01 | `WalletIsMineDescriptors`
|            8,372.88 |          119,433.26 |    0.2% |       77,389.09 |      11,220.01 |    0.1% |      0.01 | `WalletIsMineLegacy`
|            1,753.04 |          570,438.89 |    0.4% |       15,359.02 |       2,606.00 |    0.0% |      0.01 | `WalletIsMineMigratedDescriptors`