@achow101 2023-04-20
> I was looking into how to get rid of the callbacks and avoid duplicating the scriptpubkeys

I added the callback to also allow for other cases where we need to know that the scriptPubKey set has changed, such as with fast rescans with block filters.